Propagate State to ReconciliationSequenceConfig to disable cleaner reconciler on normal reconciliations (#877)
* fix(715): propagate State to ReconciliationSequenceConfig

Allows distinguishing between delete reconciliations and normal reconciliations, making sure the cleaner reconciler is only scheduled in case of delete.

Adds test case in unit test and adopts scheduler test to be status-aware.

* chore(clusterstatus): change IsDeletion > IsDeletionInProgress

This change adopts IsDeletion to be a more accurate derivative for a
cluster that is currently has a deletion in progress. It removes the
pending check as this was never used anyways

* chore: rename DesiredStatus to ReconciliationStatus

* test: change invoker clusterStateMock..Status > ClusterStatusReconciling

* test: adopt scheduler run test to have correct initial state